update styles + minor fixes
StateVarTable
 - Now displays the appropriate message depending if there are state variables or not

Graph
 - Updated buttons to use DaisyUI base button styling
 - Updated the controls on bottom left to use DaisyUI light/dark theming

Misc
 - Renamed `TestTable` to `DashGroupTable`
 - Moved `StateVarTable` to it's own file
